+++ tests/resources/demo_other/__init__.py
@@ -1,3 +1,3 @@
-#This file is generated for the module 'demo_other' on Mon Dec 14 15:54:49 2020.
+#This file is generated for the module 'demo_other' on Mon Dec 14 22:46:56 2020.
 
 from . import *